@@ -206,7 +206,7 @@ class LoginPlugin extends Plugin
 
             foreach ($pages->instances() as $page) {
                 $header = $page->header();
-                if (isset($header) && isset($header->access) && isset($header->login['visibility_requires_access']) && $header->login['visibility_requires_access'] === true) {
+                if ($header && isset($header->access) && isset($header->login['visibility_requires_access']) && $header->login['visibility_requires_access'] === true) {
                     $config = $this->mergeConfig($page);
                     $access = $this->login->isUserAuthorizedForPage($user, $page, $config);
                     if ($access === false) {
@@ -671,7 +671,6 @@ class LoginPlugin extends Plugin
         }
 
         $form->validate();
-        $form->filter();
 
         /** @var Data $form_data */
         $form_data = $form->getData();
@@ -827,7 +826,6 @@ class LoginPlugin extends Plugin
         $language = $this->grav['language'];
 
         $form->validate();
-        $form->filter();
 
         /** @var Data $form_data */
         $form_data = $form->getData();
