@@ -1,8 +1,9 @@
 <?php
 namespace Grav\Plugin;
 
+use Composer\Autoload\ClassLoader;
 use Grav\Common\Plugin;
-
+use Grav\Plugin\ShortcodeUi\ShortcodeUiTwigExtension;
 
 class ShortcodeUiPlugin extends Plugin
 {
@@ -12,10 +13,23 @@ class ShortcodeUiPlugin extends Plugin
     public static function getSubscribedEvents()
     {
         return [
+            'onPluginsInitialized' => [
+                ['autoload', 100001],
+            ],
             'onShortcodeHandlers' => ['onShortcodeHandlers', 0],
             'onTwigExtensions' => ['onTwigExtensions', 0],
             'onTwigTemplatePaths' => ['onTwigTemplatePaths', 0],
         ];
+    }
+
+    /**
+     * [onPluginsInitialized:100000] Composer autoload.
+     *
+     * @return ClassLoader
+     */
+    public function autoload()
+    {
+        return require __DIR__ . '/vendor/autoload.php';
     }
 
     /**
@@ -28,7 +42,6 @@ class ShortcodeUiPlugin extends Plugin
 
     public function onTwigExtensions()
     {
-        require_once(__DIR__ . '/twig/ShortcodeUITwigExtension.php');
         $this->grav['twig']->twig->addExtension(new ShortcodeUiTwigExtension());
     }
 
@@ -37,6 +50,6 @@ class ShortcodeUiPlugin extends Plugin
      */
     public function onShortcodeHandlers()
     {
-        $this->grav['shortcode']->registerAllShortcodes(__DIR__.'/shortcodes');
+        $this->grav['shortcode']->registerAllShortcodes(__DIR__ . '/classes/shortcodes');
     }
 }
